+++ Library/Homebrew/cmd/tap.rb
@@ -1,4 +1,5 @@
 require "tap"
+require "core_formula_repository"
 
 module Homebrew
   def tap
