+++ Library/Homebrew/formula.rb
@@ -10,6 +10,7 @@ require "software_spec"
 require "install_renamed"
 require "pkg_version"
 require "tap"
+require "core_formula_repository"
 require "formula_renames"
 require "keg"
 require "migrator"
